将库存从 int32 迁移到 int64, DLL 从 INT UNSIGNED 迁移到 BIGINT UNSIGNED

+++ internal/order/app/command/create_order.go
@@ -140,7 +140,7 @@ func (c createOrderHandler) validate(ctx context.Context, items []*domain.ItemWi
 
 // packItems 合并相同商品的数量
 func packItems(items []*domain.ItemWithQuantity) []*domain.ItemWithQuantity {
-	merged := make(map[string]int32)
+	merged := make(map[string]int64)
 	for _, item := range items {
 		merged[item.Id] += item.Quantity
 	}
